+++ app/src/main.tsx
@@ -5,13 +5,14 @@ import { createTheme, CssBaseline } from "@mui/material";
 
 import MainLayout from "./components/MainLayout";
 import PrivateRoute, { ROUTE_LOGIN } from "./components/Security/PrivateRoute";
-import HomeTabs from "./pages/Home/HomeTabs";
+import Home from "./pages/Home";
 import Login from "./pages/Login";
 import PageAlbum from "./pages/PageAlbum";
 import PageArtist from "./pages/PageArtist";
 import PageMix from "./pages/PageMix";
 import PagePlaylist from "./pages/PagePlaylist";
 import PageTrack from "./pages/PageTrack";
+import PageProcessing from "./pages/Processing";
 import Search from "./pages/Search";
 import { APIFetcherProvider } from "./provider/ApiFetcherProvider";
 import { AuthProvider } from "./provider/AuthProvider";
@@ -62,13 +63,14 @@ root.render(
                 <Route path={ROUTE_LOGIN} element={<Login />} />
                 <Route element={<PrivateRoute />}>
                   <Route element={<DefaultLayout />}>
-                    <Route path="/" element={<HomeTabs />} />
+                    <Route path="/" element={<Home />} />
                     <Route path="/search/:keywords" element={<Search />} />
                     <Route path="/artist/:id" element={<PageArtist />} />
                     <Route path="/mix/:id" element={<PageMix />} />
                     <Route path="/playlist/:id" element={<PagePlaylist />} />
                     <Route path="/album/:id" element={<PageAlbum />} />
                     <Route path="/track/:id" element={<PageTrack />} />
+                    <Route path="/processing" element={<PageProcessing />} />
                   </Route>
                 </Route>
               </Routes>